[code] [info] Render goals in the webview context.
This is a prereq for #143 as the current jsCoq render needs to run in
a full DOM-like context, which in VS Code amounts to a webview.

For that, we need to use a bundler to build the webview, we choose
`vite` which is, in our setting, basically a frontend to RollUp.

Detailed list of changes is:

- editor/code/view: new infoview location, code split from the
  extension context
- editor/code: use nested tsconfig; broken module resolution
- editor/code: use Vite to build WebView
