+++ editor/code/package.json
@@ -158,10 +158,12 @@
   "devDependencies": {
     "@types/node": "^16.11.7",
     "@types/throttle-debounce": "^5.0.0",
+    "@types/vscode-webview": "^1.57.1",
     "@types/vscode": "^1.73.0",
     "esbuild": "^0.16.13",
     "prettier": "^2.8.1",
-    "typescript": "^4.9.4"
+    "typescript": "^4.9.4",
+    "vite": "^4.0.4"
   },
   "dependencies": {
     "throttle-debounce": "^5.0.0",
@@ -173,8 +175,7 @@
     "esbuild-base": "esbuild ./src/client.ts --bundle --outfile=out/src/client.js --external:vscode --format=cjs --platform=node",
     "esbuild": "npm run esbuild-base -- --sourcemap",
     "esbuild-watch": "npm run esbuild-base -- --sourcemap --watch",
-    "compile": "tsc -p ./",
-    "test-compile": "tsc -p ./",
-    "watch": "tsc -watch -p ./"
+    "compile": "tsc -b && vite build view",
+    "watch": "tsc -watch -b"
   }
 }
