+++ editor/code/src/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../tsconfig-base.json",
+  "compilerOptions": {
+    "module": "CommonJS",
+    "target": "ESNext"
+  },
+  "include": ["**/*.ts", "../lib/**/*.ts"]
+}
