+++ editor/code/tsconfig.json
@@ -1,13 +1,4 @@
 {
-  "compilerOptions": {
-    "module": "commonjs",
-    "target": "es2020",
-    "outDir": "out",
-    "sourceMap": true,
-    "strict": true,
-    "rootDir": "."
-  },
-  "files": ["src/client.ts"],
-  // Default if not specified is node_modules, out, etc...
-  "exclude": ["node_modules", "out", ".vscode-test"]
+  "files": [],
+  "references": [{ "path": "src" }, { "path": "view" }]
 }
