+++ editor/code/view/index.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <script src="infoview.ts" type="module"></script>
+    <title>Coq's info panel</title>
+  </head>
+  <body>
+    <div id="coq-info-body"></div>
+  </body>
+</html>
