+++ editor/code/view/vite.config.mts
@@ -1,0 +1,15 @@
+import { defineConfig } from "vite";
+
+export default defineConfig({
+  build: {
+    outDir: "out",
+    sourcemap: "inline",
+    rollupOptions: {
+      output: {
+        entryFileNames: `[name].js`,
+        chunkFileNames: `[name].js`,
+        assetFileNames: `[name].[ext]`,
+      },
+    },
+  },
+});
